# Implementation pack/05-openapi.yaml
openapi: 3.0.3
info:
  title: AI CV Tailoring and Cover Letter Platform — Backend API
  description: >
    Draft specification for developer implementation. Covers CV ingestion,
    job post ingestion, matching, tailored CV generation, guided cover
    letter generation, and export. All generated content must carry
    evidence references — see evidenceReferences fields throughout.
  version: 0.1.0-draft
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /matches:
    post:
      tags: [matching]
      summary: Create a match analysis between a CV profile and a job post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cvId, jobPostId]
              properties:
                cvId: { type: string, format: uuid }
                jobPostId: { type: string, format: uuid }
      responses:
        '202':
          description: Match analysis queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProcessingJobRef' }
    get:
      tags: [matching]
      summary: List match runs for the current user, most recent first
      description: >-
        Account-only (matches GET /cvs and GET /job-posts' precedent) — a
        trial session can already see its own single match via GET
        /matches/{matchId}, but a browsable list is an account feature.
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: Paginated list of matches
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MatchListResponse' }
  /matches/{matchId}:
    parameters:
      - $ref: '#/components/parameters/matchId'
    get:
      tags: [matching]
      summary: Match analysis and evidence flags
      responses:
        '200':
          description: Match result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MatchRun' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    matchId:
      name: matchId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
    offset:
      name: offset
      in: query
      required: false
      schema: { type: integer, default: 0, minimum: 0 }
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      description: >
        Standard error envelope returned by every non-2xx response.
        `details` is free-form and endpoint-specific — populate it with
        whatever helps the caller correct the request (offending field,
        allowed values, etc.), omit it when there's nothing further to add.
      required: [status, code, message, timestamp, path]
      properties:
        status: { type: integer, description: "HTTP status code, duplicated in the body for logging convenience" }
        code: { type: string, description: "Stable machine-readable error code, e.g. VALIDATION_ERROR" }
        message: { type: string, description: "Human-readable explanation" }
        timestamp: { type: string, format: date-time }
        path: { type: string, description: "Request path that produced the error" }
        details:
          type: object
          additionalProperties: true
          nullable: true
      example:
        status: 400
        code: VALIDATION_ERROR
        message: "Invalid file type. Only PDF and DOCX are supported."
        timestamp: "2026-08-03T10:00:00Z"
        path: "/api/v1/cvs"
        details:
          field: file
          reason: unsupported_mime_type
          allowed: ["application/pdf", "application/vnd.openxmlformats-officedocument.wordprocessingml.document"]
...
    MatchRun:
      type: object
      properties:
        id: { type: string, format: uuid }
        cvProfileVersionId: { type: string, format: uuid }
        jobPostProfileId: { type: string, format: uuid }
        status: { type: string, enum: [queued, processing, completed, failed] }
        summaryAnalysis: { type: object, additionalProperties: true }
        unsupportedRequirementCount: { type: integer }
        evidenceItems:
          type: array
          items: { $ref: '#/components/schemas/MatchEvidenceItem' }
        createdAt: { type: string, format: date-time }
      example:
        id: "9a8b7c6d-5e4f-4321-8888-aabbccddeeff"
        cvProfileVersionId: "5566aabb-1122-4c33-9d44-eeff00112233"
        jobPostProfileId: "1a2b3c4d-5e6f-4788-9900-112233445566"
        status: completed
        summaryAnalysis: { totalRequirements: 8, supported: 5, partiallySupported: 2, unsupported: 1, score: 75.0 }
        unsupportedRequirementCount: 1
        evidenceItems:
          - id: "req_1"
            requirementText: "5+ years Python experience"
            requirementType: required
            supportLevel: supported
            sourceReferences: ["exp_1"]
          - id: "req_4"
            requirementText: "Leadership experience"
            requirementType: required
            supportLevel: partially_supported
            sourceReferences: ["exp_1"]
          - id: "req_5"
            requirementText: "Terraform experience"
            requirementType: preferred
            supportLevel: unsupported
            sourceReferences: []
        createdAt: "2026-08-03T10:30:00Z"
    MatchListItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        jobPostId: { type: string, format: uuid }
        jobTitle: { type: string, nullable: true }
        employer: { type: string, nullable: true }
        status: { type: string, enum: [queued, processing, completed, failed] }
        score: { type: number, nullable: true }
        createdAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }
    MatchListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/MatchListItem' }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
    MatchEvidenceItem:
      type: object
      description: >
        supportLevel: unsupported, contradictory, and unclear are three
        distinct failure modes, not variants of one "not usable" state —
        see 03-data-model.md section on match_evidence_items for the full
        definition of each and what's permitted in generated output.
        None of the three ever trigger fabrication upstream: each is
        surfaced to the user (directly, or via a cover letter question)
        rather than papered over in generation.
      properties:
        id: { type: string, format: uuid }
        requirementText: { type: string }
        requirementType: { type: string, enum: [required, preferred] }
        supportLevel: { type: string, enum: [supported, partially_supported, unsupported, contradictory, unclear] }
        sourceReferences:
          type: array
          items: { type: string, description: "Reference to a cv_experience_items.id, cv_skill_items.id, cv_education_items.id, or cover_letter_answers.id. Empty array only valid when supportLevel is unsupported or unclear. For contradictory, holds the conflicting sources themselves." }
...
    ProcessingJobRef:
      type: object
      properties:
        jobId: { type: string, format: uuid }
        status: { type: string, enum: [queued] }
...